move message_utils.hpp to util folder add proper namespace to source files
