+++ include/spine/command_factory.h
@@ -53,11 +53,6 @@ public:
     CommandPtr create(const Message& message, int target_version);
 
 private:
-    //TODO: make common
-    Any* to_any(const Message& message);
-    Any* to_any(const Message& message, const std::string& type_url_prefix);
-
-private:
     std::unique_ptr<ActorContext> actor_context_;
     Poco::UUIDGenerator uuid_generator_;
